refactor: change directories names
