+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import { Layout } from './components/Layout/Layout';
-import { HomePage } from './components/HomePageComponent/HomePage';
-import { CatsPage } from './components/CatsPageComponent/CatsPage';
-import { AboutPage } from './components/AboutPageComponent/AboutPage';
-import { NotFoundPage } from './components/NotFoundPageComponent/NotFoundPage';
+import { HomePage } from './components/HomePage/HomePage';
+import { CatsPage } from './components/CatsPage/CatsPage';
+import { AboutPage } from './components/AboutPage/AboutPage';
+import { NotFoundPage } from './components/NotFoundPage/NotFoundPage';
 
 export const App = () => {
   return (
