+++ src/components/basicComponents/Inputs/RadioInput.tsx
@@ -1,5 +1,5 @@
 import { FieldErrors, FieldValues, Path, UseFormRegister, WatchInternal } from 'react-hook-form';
-import { IFormValues } from '../../CatsPageComponent/CatsPage';
+import { IFormValues } from '../../CatsPage/CatsPage';
 import './RadioInput.css';
 import { Tooltip } from '../Tooltip';
