+++ src/components/basicComponents/Inputs/RangeInput.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useEffect, useState } from 'react';
 import { FieldValues, Path, UseFormRegister } from 'react-hook-form';
-import { IFormValues } from '../../CatsPageComponent/CatsPage';
+import { IFormValues } from '../../CatsPage/CatsPage';
 import './RangeInput.css';
 import { Rating } from '../Rating';
