+++ src/components/basicComponents/Inputs/SelectInput.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useState } from 'react';
 import { FieldErrors, FieldValues, Path, UseFormRegister } from 'react-hook-form';
-import { IFormValues } from '../../CatsPageComponent/CatsPage';
+import { IFormValues } from '../../CatsPage/CatsPage';
 import './SelectInput.css';
 import { Tooltip } from '../Tooltip';
